04-threat-intel-dashboard/requirements.txt
anthropic
requests
python-dotenv
OTXv2